+++ src/textual/css/constants.py
@@ -90,6 +90,7 @@ VALID_HATCH: Final = {"left", "right", "cross", "vertical", "horizontal"}
 VALID_TEXT_WRAP: Final = {"wrap", "nowrap"}
 VALID_TEXT_OVERFLOW: Final = {"clip", "fold", "ellipsis"}
 VALID_EXPAND: Final = {"greedy", "optimal"}
+VALID_SCROLLBAR_VISIBILITY: Final = {"visible", "hidden"}
 
 HATCHES: Final = {
     "left": "╲",